refactor: split app into components
